chore(bench/node): split suites and add benchmark for parallel-js only
