+++ packages/bench/benches/par.js
@@ -1,5 +1,3 @@
-import nodeUtil from 'node:util'
-import { expandSuitesWithDerived, suites } from '../src/suites/index.js'
 import * as bencher from '../src/bencher.js'
 import {
   getRolldownSuiteList,
@@ -7,12 +5,10 @@ import {
   runRolldown,
   runRollup,
 } from '../src/run-bundler.js'
+import { suiteRomeTs } from '../src/suites/rome-ts.js'
+import { expandSuitesWithDerived } from '../src/suites/index.js'
 
-console.log(
-  nodeUtil.inspect(suites, { depth: null, colors: true, showHidden: false }),
-)
-
-for (const suite of expandSuitesWithDerived(suites)) {
+for (const suite of expandSuitesWithDerived([suiteRomeTs])) {
   const excludedBundlers = Array.isArray(suite.disableBundler)
     ? suite.disableBundler
     : suite.disableBundler
