+++ packages/bench/src/utils.js
@@ -10,3 +10,12 @@ const dirname = nodePath.dirname(nodeUrl.fileURLToPath(import.meta.url))
 export const REPO_ROOT = nodePath.join(dirname, '../../..')
 
 export const PROJECT_ROOT = nodePath.join(dirname, '..')
+
+/**
+ *
+ * @param {import('./types.js').BenchSuite} config
+ * @returns {import('./types.js').BenchSuite}
+ */
+export function defineSuite(config) {
+  return config
+}
